use k8s cluster default imagePullPolicy
K8s sets "IfNotPresent" image pull policy by default when it is not
explicitly set. As well as it sets the image pull policy to "Always" for
the images with ":latest" tag or the ones that do not specify the tag.

The intention is to keep this default and widely accepted behavior.

This also avoids surprises as when one would deploy, say "nginx:latest",
he will get the latest patched nginx instead of an old cached by the
provider image full of CVE's.